+++ fiddles/java/fiddle-0007-ChromeDriverHW/chromedriverhw/src/main/java/browser/Browser.java
@@ -1,0 +1,11 @@
+package browser;
+
+import java.net.MalformedURLException;
+
+public interface Browser {
+
+  String start() throws InterruptedException, MalformedURLException;
+
+  void quit();
+
+}